+++ CHANGELOG.md
@@ -8,5 +8,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Add site_builder_console:bundle:create command
+- Add site_builder_console:bundle:delete command
 
 [Unreleased]: https://github.com/wongjn/site_builder_console/commits